third_party/blink/web_tests/platform/generic/fast/dom/constants-expected.txt

This page tests CSSRule, NodeFilter, and Event. It tests:

Whether their global constructors have the correct constant values
Whether their objects have the correct constant values, except NodeFilter
PASS: cssRule.STYLE_RULE should be 1 and is.
PASS: cssRule.CHARSET_RULE should be 2 and is.
PASS: cssRule.IMPORT_RULE should be 3 and is.
PASS: cssRule.MEDIA_RULE should be 4 and is.
PASS: cssRule.FONT_FACE_RULE should be 5 and is.
PASS: cssRule.PAGE_RULE should be 6 and is.
PASS: window.CSSRule.STYLE_RULE should be 1 and is.
PASS: window.CSSRule.CHARSET_RULE should be 2 and is.
PASS: window.CSSRule.IMPORT_RULE should be 3 and is.
PASS: window.CSSRule.MEDIA_RULE should be 4 and is.
PASS: window.CSSRule.FONT_FACE_RULE should be 5 and is.
PASS: window.CSSRule.PAGE_RULE should be 6 and is.
PASS: window.NodeFilter.FILTER_ACCEPT should be 1 and is.
PASS: window.NodeFilter.FILTER_REJECT should be 2 and is.
PASS: window.NodeFilter.FILTER_SKIP should be 3 and is.
PASS: window.NodeFilter.SHOW_ALL should be 4294967295 and is.
PASS: window.NodeFilter.SHOW_ELEMENT should be 1 and is.
PASS: window.NodeFilter.SHOW_ATTRIBUTE should be 2 and is.
PASS: window.NodeFilter.SHOW_TEXT should be 4 and is.
PASS: window.NodeFilter.SHOW_CDATA_SECTION should be 8 and is.
PASS: window.NodeFilter.SHOW_ENTITY_REFERENCE should be 16 and is.
PASS: window.NodeFilter.SHOW_ENTITY should be 32 and is.
PASS: window.NodeFilter.SHOW_PROCESSING_INSTRUCTION should be 64 and is.
PASS: window.NodeFilter.SHOW_COMMENT should be 128 and is.
PASS: window.NodeFilter.SHOW_DOCUMENT should be 256 and is.
PASS: window.NodeFilter.SHOW_DOCUMENT_TYPE should be 512 and is.
PASS: window.NodeFilter.SHOW_DOCUMENT_FRAGMENT should be 1024 and is.
PASS: window.NodeFilter.SHOW_NOTATION should be 2048 and is.
PASS: event.NONE should be 0 and is.
PASS: event.CAPTURING_PHASE should be 1 and is.
PASS: event.AT_TARGET should be 2 and is.
PASS: event.BUBBLING_PHASE should be 3 and is.
PASS: window.Event.NONE should be 0 and is.
PASS: window.Event.CAPTURING_PHASE should be 1 and is.
PASS: window.Event.AT_TARGET should be 2 and is.
PASS: window.Event.BUBBLING_PHASE should be 3 and is.
